Added store buttonTitle support and broke out framework shapes.
