+++ how-to/customize-workspace/client/src/framework/shapes/bootstrap-shapes.ts
@@ -1,0 +1,9 @@
+export type BootstrapComponents = "home" | "store" | "dock" | "none";
+
+export interface BootstrapOptions {
+	store: boolean;
+	home: boolean;
+	dock: boolean;
+	notifications: boolean;
+	autoShow: BootstrapComponents[];
+}
