+++ how-to/customize-workspace/client/src/framework/shapes/headless-shapes.ts
@@ -1,0 +1,4 @@
+import type { ModuleList } from "./module-shapes";
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface HeadlessProviderOptions extends ModuleList {}
