+++ how-to/customize-workspace/client/src/framework/shapes/home-shapes.ts
@@ -1,0 +1,9 @@
+export interface HomeProviderOptions {
+	id: string;
+	title: string;
+	icon: string;
+	queryMinLength?: number;
+	queryAgainst?: string[];
+	enablePageIntegration?: boolean;
+	enableWorkspaceIntegration?: boolean;
+}
