+++ how-to/customize-workspace/client/src/framework/shapes/index.ts
@@ -3,10 +3,14 @@ export * from "./auth-shapes";
 export * from "./conditions-shapes";
 export * from "./connection-shapes";
 export * from "./endpoint-shapes";
-export * from "./framework-shapes";
+export * from "./setting-shapes";
 export * from "./init-options-shapes";
 export * from "./integrations-shapes";
 export * from "./lifecycle-shapes";
 export * from "./logger-shapes";
 export * from "./module-shapes";
 export * from "./template-shapes";
+export * from "./store-shapes";
+export * from "./bootstrap-shapes";
+export * from "./browser-shapes";
+export * from "./app-shapes";
