+++ how-to/customize-workspace/client/src/framework/shapes/notification-shapes.ts
@@ -1,0 +1,4 @@
+import type { NotificationsPlatform } from "@openfin/workspace/notifications";
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface NotificationProviderOptions extends NotificationsPlatform {}
