+++ how-to/customize-workspace/client/src/framework/shapes/platform-shapes.ts
@@ -1,0 +1,9 @@
+export interface PlatformProviderOptions {
+	rootUrl: string;
+	sharing: boolean;
+	intentPicker?: {
+		url: string;
+		height?: number;
+		width?: number;
+	};
+}
