+++ how-to/customize-workspace/client/src/framework/shapes/theme-shapes.ts
@@ -1,0 +1,5 @@
+import type { CustomThemes } from "@openfin/workspace-platform";
+
+export interface ThemeProviderOptions {
+	themes: CustomThemes;
+}
