+++ how-to/customize-workspace/public/manifest.fin.json
@@ -491,7 +491,8 @@
 							"image": {
 								"src": "http://localhost:8080/common/images/coding-1-unsplash.jpg"
 							},
-							"tags": ["expero"]
+							"tags": ["expero"],
+							"buttonTitle": "View"
 						},
 						{
 							"title": "Dev Tools",
@@ -499,7 +500,8 @@
 							"image": {
 								"src": "http://localhost:8080/common/images/coding-2-unsplash.jpg"
 							},
-							"tags": ["tools"]
+							"tags": ["tools"],
+							"buttonTitle": "View"
 						},
 						{
 							"title": "Learning Resource",
@@ -507,7 +509,8 @@
 							"image": {
 								"src": "http://localhost:8080/common/images/coding-3-unsplash.jpg"
 							},
-							"tags": ["page"]
+							"tags": ["page"],
+							"buttonTitle": "View"
 						}
 					]
 				},
@@ -524,7 +527,8 @@
 							"image": {
 								"src": "http://localhost:8080/common/images/coding-4-unsplash.jpg"
 							},
-							"tags": ["view"]
+							"tags": ["view"],
+							"buttonTitle": "View"
 						},
 						{
 							"title": "Web Apps",
@@ -532,7 +536,8 @@
 							"image": {
 								"src": "http://localhost:8080/common/images/coding-5-unsplash.jpg"
 							},
-							"tags": ["manifest"]
+							"tags": ["manifest"],
+							"buttonTitle": "View"
 						},
 						{
 							"title": "Native Apps",
@@ -540,7 +545,8 @@
 							"image": {
 								"src": "http://localhost:8080/common/images/coding-6-unsplash.jpg"
 							},
-							"tags": ["native"]
+							"tags": ["native"],
+							"buttonTitle": "View"
 						}
 					]
 				}
