+++ how-to/customize-workspace/public/settings.json
@@ -570,7 +570,8 @@
 						"image": {
 							"src": "http://localhost:8080/common/images/coding-1-unsplash.jpg"
 						},
-						"tags": ["expero"]
+						"tags": ["expero"],
+						"buttonTitle": "View"
 					},
 					{
 						"title": "Dev Tools",
@@ -578,7 +579,8 @@
 						"image": {
 							"src": "http://localhost:8080/common/images/coding-2-unsplash.jpg"
 						},
-						"tags": ["tools"]
+						"tags": ["tools"],
+						"buttonTitle": "View"
 					},
 					{
 						"title": "Learning Resource",
@@ -586,7 +588,8 @@
 						"image": {
 							"src": "http://localhost:8080/common/images/coding-3-unsplash.jpg"
 						},
-						"tags": ["page"]
+						"tags": ["page"],
+						"buttonTitle": "View"
 					}
 				]
 			},
@@ -603,7 +606,8 @@
 						"image": {
 							"src": "http://localhost:8080/common/images/coding-4-unsplash.jpg"
 						},
-						"tags": ["view"]
+						"tags": ["view"],
+						"buttonTitle": "View"
 					},
 					{
 						"title": "Web Apps",
@@ -611,7 +615,8 @@
 						"image": {
 							"src": "http://localhost:8080/common/images/coding-5-unsplash.jpg"
 						},
-						"tags": ["manifest"]
+						"tags": ["manifest"],
+						"buttonTitle": "View"
 					},
 					{
 						"title": "Native Apps",
@@ -619,7 +624,8 @@
 						"image": {
 							"src": "http://localhost:8080/common/images/coding-6-unsplash.jpg"
 						},
-						"tags": ["native"]
+						"tags": ["native"],
+						"buttonTitle": "View"
 					}
 				]
 			}
